Restore old default behavior of Ifpack
4C used to hard-code certain Ifpack parameters to overwrite Ifpack's
default behavior. This commit restores the old 4C settings by including
them into the Ifpack xml file. This should bring us back to the old way
of running Ifpack.

If this change makes the pipeline pass, we still can decide whether we
want to recommend these settings to the users or if we want to rely on
Ifpack's default values (yet with slight changes in 4C test results).